Reworked some of the internals on MineResets to prepare for the next phase of work.  Created a MinTasks class to house some of the tasks that are more task oriented.
